refactor: update AnimalController to support batch animal addition and modify Animal entity methods

+++ src/main/java/com/example/demo/service/AnimalService.java
@@ -40,6 +40,6 @@ public class AnimalService {
     }
 
     public Animal getAnimalById(Long id) {
-        return animalRepository.getAnimalById(id).get();
+        return animalRepository.getAnimalById(id.longValue()).get();
     }
 }
